Move `merlin.core.worker` to `merlin.io.worker`
Resolves #4
